+++ scripts/generate.ts
@@ -17,11 +17,37 @@
  */
 
 import {existsSync} from "node:fs"
-import {RawConfigSchema} from "../app/config.ts"
-import type {Config as ServerConfig} from "../app/server.ts"
-import {Server} from "../app/server.ts"
-import type {ListToolsResult} from "@modelcontextprotocol/sdk/types.js"
 import {readFile, writeFile} from "node:fs/promises"
+import type {ListToolsResult} from "@modelcontextprotocol/sdk/types.js"
+import {RawConfigSchema} from "../app/config.ts"
+import type {Config as ServerConfig} from "../lib/server.ts"
+import {Server} from "../lib/server.ts"
+
+/**
+ * {@link https://code.visualstudio.com/docs/reference/variables-reference/#_input-variables | VS Code Reference}
+ */
+interface VscodeInput {
+	type: string
+	id: string
+	description: string
+	password?: boolean
+}
+
+/**
+ * {@link https://code.visualstudio.com/docs/copilot/chat/mcp-servers/#_configuration-format | VS Code Reference}
+ */
+interface VscodeConfig {
+	env: Record<string, string>
+	command: string
+	args: string[]
+}
+
+interface VscodeQuery {
+	name: string
+	inputs: string
+	config: string
+	quality?: string
+}
 
 async function main(): Promise<void> {
 	if (!existsSync("README.md")) {
@@ -43,14 +69,67 @@ async function main(): Promise<void> {
 		return a.name.localeCompare(b.name)
 	})
 
+	let badges = createBadges(RawConfigSchema.shape)
 	let config = formatConfig(RawConfigSchema.shape)
 	let tools = formatTools(ls)
 
 	let output = input
+	output = insert("badges", output, badges)
 	output = insert("config", output, config)
 	output = insert("tools", output, tools)
 
 	await writeFile("README.md", output, "utf8")
+}
+
+function createBadges(shape: typeof RawConfigSchema.shape): string {
+	let bru = "https://badgen.net/static/Open%20in%20VS%20Code/npx/blue"
+	let biu = "https://badgen.net/static/Open%20in%20VS%20Code%20Insiders/npx/cyan"
+
+	let inputs: VscodeInput[] = [
+		{
+			type: "promptString",
+			id: "docspace_base_url",
+			description: String(shape.DOCSPACE_BASE_URL.description).replaceAll("`", ""),
+		},
+		{
+			type: "promptString",
+			id: "docspace_api_key",
+			description: String(shape.DOCSPACE_API_KEY.description).replaceAll("`", ""),
+			password: true,
+		},
+	]
+
+	let config: VscodeConfig = {
+		env: {
+			DOCSPACE_BASE_URL: "${input:docspace_base_url}",
+			DOCSPACE_API_KEY: "${input:docspace_api_key}",
+		},
+		command: "npx",
+		args: ["--yes", "@onlyoffice/docspace-mcp"],
+	}
+
+	let query: VscodeQuery = {
+		name: "onlyoffice-docspace",
+		inputs: JSON.stringify(inputs),
+		config: JSON.stringify(config),
+	}
+
+	let vbu = "https://insiders.vscode.dev/redirect/mcp/install"
+
+	let vru = new URL(vbu)
+	vru.search = new URLSearchParams({...query}).toString()
+
+	query.quality = "insiders"
+
+	let viu = new URL(vbu)
+	viu.search = new URLSearchParams({...query}).toString()
+
+	let c = ""
+
+	c += `[![Open in VS Code using npx command](${bru})](${vru})\n`
+	c += `[![Open in VS Code Insiders using npx command](${biu})](${viu})`
+
+	return c
 }
 
 function formatConfig(shape: typeof RawConfigSchema.shape): string {